Make TLS settings actually usable in BOSH manifests
Replace configurable paths with hard-coded (jobs-dir relative) paths and
outfit the job spec with new properties for the _contents_ of the
certificate and key files.

The vault.listener.tcp.tlsdisable boolean is now gone, in favor of the
presence / absence of the certificate and key properties to turn TLS on
or off.